+++ src/block.ts
@@ -183,7 +183,9 @@ export abstract class Block
 
         for (let index = 0; index < contextArray.length; index++) {
             let char = contextArray[index];
-            if (char === divider && opened === 0) {
+            if (char === divider && index === contextArray.length - 1) {
+                break;
+            } else if (char === divider && opened === 0) {
                 endPos = index;
                 result.push(context.substr(startPos, endPos - startPos));
                 startPos = index + 1;